sort_test: Fix issue where data isn't sorted
In the previous implementation, when testing with random numbers
generated by the xoroshiro128p PRNG, there was a chance that the kernel
device module would stop because it couldn't pass the `check_list()`
function.

The main reason was that the pseudorandom number generated by
xoroshiro128p is of the `uint64_t` data type (unsigned), while the data
type used for storing data in the element structure for the linked-list
is `int` (signed). To resolve this, I made changes to scale down the
generated number by a specific constant and then implicitly assign the
result to the `int` data type, avoiding the probability of integer
overflow.

Additionally, I modified the sort test in the kernel device module to
perform a single sort test, rather than a loop sort test. The loop sort
test will be applied in the user space program `client.c`.