client: Apply test with continuous nodes
This version of the client introduces testing with continuous nodes and
a file output function to save the results of the kernel sorting test
for specific sorting programs. Additionally, the implementation of the
single node number test has been modified to apply different sorting
programs, similar to the continuous node test.

For the continuous node test, it is designed to test the sorting
programs with different numbers of nodes. For each node count, the
sorting programs will be tested with 10 loops.

However, with the current implementation of the kernel device module and
user space client program, triggering the test causes severe lag on the
device, leading to crashes. The exact segment of code in the kernel
device module causing this issue is unclear. Further testing and more
accurate logging techniques are needed to identify the source of the
problem.