sort_test: New strategy for data passing & bug fix
This commit introduces a new strategy for data passing between user and
kernel space and fixes a critical issue in the kernel device for sorting
performance tests.

Firstly, I changed the data transfer method between user and kernel
space to an older strategy I used previously[1], using character buffer
strings. The buffer is initialized in a 512-size character array,
containing the test information and results. The information in the
buffer is tokenized in a specific format using the string tokenizer in
the `string.h` library in both user and kernel space.

Additionally, I fixed a bug in the kernel device module that caused the
device to crash during sorting performance tests. Previously, I
allocated kernel memory for the `element_t` list structure nodes using
the `kmalloc` function, which led to illegal memory allocation and
deallocation, causing system crashes. I changed the memory allocation
to be done individually for each node instead of all at once. After
applying this change, the issue of kernel sorting performance tests
crashing the system is resolved. On top of that, I updated the
`copy_list()` function to use the same memory allocation method as the
sampling list.

Ref:
 [1] os_2022_hw2-randyuncle-main
 https://github.com/randyuncle/My-personal-bachelar-coding-experience/blob/main/Operating%20System/os_2022_hw2-randyuncle-main/My_proc.c