sort_test: Update `client` and sort programs
With the previous changes in the data passing strategies between user
and kernel spaces, I replaced the data sending and receiving technique
with character buffer strings. Additionally, I modified the user space
program to receive arguments that determine whether to perform the
continuous amount of nodes performance test or the single amount of
nodes performance test. Last but not least, I implemented the k value
calculation for the number of comparisons as introduced in the paper
"Bottom-Up Mergesort - A Detailed Analysis"[1].

On the other hand, for the sort programs, I replaced "list.h" with
<linux/list.h>, the Linux kernel internal list library.

Ref:
 [1] Bottom-Up Mergesort - A Detailed Analysis
 https://link.springer.com/article/10.1007/BF01294131