sort_test: Fix some typos
This commit fixes the typos in the words "continuous" and
"continuously", which were misspelled as "continious" and "continuesly"
in `client.c` and the Makefile.

Additionally, in `sort_test_kernel.c`, "interrupt" was misspelled as
"interupt".

Moreover, I expanded the comment for the use of the <linux/uaccess.h>
library in `sort_test_kernel.c`. It originally stated
"for copy_from_user", but both copy_from_user() and copy_to_user()
functions are from that library.