sort_test: Comment out some printk operations
In the previous version of the kernel device driver, every time it was
opened, closed, or written to, it would output messages using printk()
to dmesg, the kernel log. However, since the `./client continuous`
command applies many tests, this could easily fill up the memory space
in the root file system, making the tests unable to perform. To prevent
this issue, I commented out the printk() calls that do not print error
messages in the kernel device driver.