sort: Introduce alpha-merge sort
In paper "Strategies for Stable Merge Sorting", the authors introduces a
natural merge sort algorithm -- alpha-merge sort, where alpha is defined
as `golden ratio < alpha <= 2`. This sort aims to shorten the distance
between upper bound and lower bound of the merge cost in Timsort. With
the experiment from the original author, they found out that when alpha
sets to 1.62, it shares a stable upper and lower bound of the normalized
merge cost.

With the goal of know the effectness of the alpha-merge sort, I
introduce it into this repository, and make it be compilable in the
kernel space.